+++ src/Services/RemoteEventsFactory.php
@@ -26,6 +26,7 @@ use Bitrix24\SDK\Services\CRM\Company\Events\CrmCompanyEventsFactory;
 use Bitrix24\SDK\Services\CRM\Contact\Events\CrmContactEventsFactory;
 use Bitrix24\SDK\Services\SonetGroup\Events\SonetGroupEventsFactory;
 use Bitrix24\SDK\Services\Telephony\Events\TelephonyEventsFactory;
+use Bitrix24\SDK\Services\IMOpenLines\Connector\Events\ImConnectorEventsFactory;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -223,6 +224,7 @@ readonly class RemoteEventsFactory
                 new CrmContactEventsFactory(),
                 new SonetGroupEventsFactory(),
                 new Sale\Events\SaleEventsFactory(),
+                new ImConnectorEventsFactory(),
             ],
             $logger
         );