+++ tests/ApplicationBridge/ApplicationCredentialsProvider.php
@@ -17,6 +17,8 @@ namespace Bitrix24\SDK\Tests\ApplicationBridge;
 use Bitrix24\SDK\Core\Credentials\AuthToken;
 use Bitrix24\SDK\Core\Credentials\ApplicationProfile;
 use Bitrix24\SDK\Core\Credentials\Credentials;
+use Bitrix24\SDK\Core\Credentials\DefaultOAuthServerUrl;
+use Bitrix24\SDK\Core\Credentials\Endpoints;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Bitrix24\SDK\Events\AuthTokenRenewedEvent;
 use JetBrains\PhpStorm\NoReturn;
@@ -47,7 +49,7 @@ readonly class ApplicationCredentialsProvider
             null,
             $this->repository->getToken(),
             $applicationProfile,
-            $domainUrl
+            new Endpoints($domainUrl, DefaultOAuthServerUrl::default())
         );
     }